added functions for plotting deformed shapes
Two functions added:
recordNodeDisp(file) saves the displacements of all nodes in a file.
plot_deformedshape(file, tstep, scale) takes a file such as the exported by recordNodeDisp and plots the deformed shape at the specified step of the analysis (tstep, first step is tstep = 1). It also returns the  figure in a variable in case the user wants to manipulate it.